Refactored the excel function and added chronological ordering in combined excel
