+++ backend/modules/services/excel_helper_modules/sanitization.py
@@ -1,0 +1,5 @@
+import re
+
+def sanitize_sheet_name(sheet_name):
+    """Sanitizes sheet names to meet Excel restrictions."""
+    return re.sub(r'[\\/*?:\[\]]', '', sheet_name)[:31]
